Secondary

+++ 10. Exams/4th/src/main/java/softuni/exam/instagraphlite/repository/PostRepository.java
@@ -2,7 +2,10 @@ package softuni.exam.instagraphlite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import softuni.exam.instagraphlite.models.entities.Post;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Integer> {
+
+
 }
