+++ 10. Exams/4th/src/main/java/softuni/exam/instagraphlite/repository/UserRepository.java
@@ -2,7 +2,9 @@ package softuni.exam.instagraphlite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import softuni.exam.instagraphlite.models.entities.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+
 }
